Remove sample-ios-app project and build final binary for iOS in sample-mpp-module
