+++ sample-ios-app/.gitignore
