--- sample-ios-app/build.gradle
@@ -1,31 +1,0 @@
-plugins {
-    id 'kotlin-multiplatform'
-}
-
-kotlin {
-    targets {
-        // fromPreset(presets.iosArm64, 'iosCommon')
-        fromPreset(presets.iosArm64, 'ios64') {
-            binaries.framework('ReactiveSample') {
-                export project(':sample-mpp-module')
-                transitiveExport = true
-            }
-        }
-        fromPreset(presets.iosX64, 'iosSim') {
-            binaries.framework('ReactiveSample') {
-                export project(':sample-mpp-module')
-                transitiveExport = true
-            }
-        }
-    }
-    sourceSets {
-        iosCommonMain {
-            dependencies {
-                api 'org.jetbrains.kotlin:kotlin-stdlib-common'
-                api project(':sample-mpp-module')
-            }
-        }
-        ios64Main.dependsOn iosCommonMain
-        iosSimMain.dependsOn iosCommonMain
-    }
-}
