+++ sample-ios-app/sample-ios-app.xcodeproj/project.xcworkspace/contents.xcworkspacedata
