+++ sample-ios-app/sample-ios-app.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
