+++ sample-ios-app/sample-ios-app.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
