+++ sample-ios-app/sample-ios-app/Assets.xcassets/AppIcon.appiconset/Contents.json
