+++ sample-ios-app/sample-ios-app/Assets.xcassets/Contents.json
