+++ sample-ios-app/sample-ios-app/Base.lproj/LaunchScreen.storyboard
