+++ sample-ios-app/sample-ios-app/ViewController.swift
@@ -1,5 +1,5 @@
 import UIKit
-import ReactiveSample
+import sample_mpp_module
 import Alamofire
 import AlamofireImage
