--- sample-ios-app/src/iosCommonMain/kotlin/com.badoo.reaktive.sample.ios/ReaktiveSample.kt
@@ -1,6 +1,0 @@
-package com.badoo.reaktive.sample.ios
-
-/**
- * It is empty interface to not skip compilation of this module with "NO-SOURCE" result.
- */
-interface ReaktiveSample
