--- sample-ios-app/xcode/sample-ios-app/KittenDataSourceImpl.swift
@@ -1,16 +1,0 @@
-import Foundation
-import ReactiveSample
-import Alamofire
-
-class KittenDataSourceImpl: KittenDataSource {
-    func load(url: String, continuation: @escaping (String?, KotlinThrowable?) -> KotlinUnit) {
-        AF.request(url)
-            .responseString { (response: DataResponse<String>) in
-                do {
-                    try continuation(response.result.get(), nil)
-                } catch {
-                    continuation(nil, KotlinRuntimeException(message: error.localizedDescription))
-                }
-        }
-    }
-}
