+++ sample-mpp-module/build.gradle
@@ -3,27 +3,40 @@
  */
 
 setupMultiplatformLibrary(project)
+setupAllTargetsWithDefaultSourceSets(project)
 
 kotlin {
-    linuxX64("linuxX64") {
-        compilations.main.cinterops {
-            libcurl.includeDirs.headerFilterOnly '/usr/include', '/usr/include/x86_64-linux-gnu'
-        }
-    }
-
     targets {
-//        fromPreset(presets.iosX64, 'iosCommon')
+        if (Target.currentTarget().isCommon()) {
+            linuxX64 {
+                compilations.main.cinterops {
+                    libcurl.includeDirs.headerFilterOnly '/usr/include', '/usr/include/x86_64-linux-gnu'
+                }
+            }
+        }
+        if (Target.currentTarget().isIos()) {
+            [ios32, ios64, iosSim].each {
+                [it.binaries.getFramework('DEBUG'), it.binaries.getFramework('RELEASE')].each {
+                    // Reaktive should be included into framework file, so we don't need to generate 2 framework files
+                    it.export project(':reaktive')
+                    it.transitiveExport = true
+                }
+            }
+        }
     }
 
     sourceSets {
         commonMain {
             dependencies {
-                implementation kotlin('test-common')
                 implementation project(':reaktive')
                 implementation project(':reaktive-annotations')
             }
         }
+        iosCommonMain {
+            dependencies {
+                // Reaktive should be included into framework file, so we don't need to generate 2 framework files
+                api project(':reaktive')
+            }
+        }
     }
 }
-
-setupAllTargetsWithDefaultSourceSets(project)
