+++ sample-mpp-module/src/iosCommonMain/kotlin/com/badoo/reaktive/samplemppmodule/KittenDataSource.kt
@@ -2,5 +2,10 @@ package com.badoo.reaktive.samplemppmodule
 
 interface KittenDataSource {
 
-    fun load(url: String, continuation: (String?, Throwable?) -> Unit)
+    fun load(url: String, continuation: (Result) -> Unit)
+
+    sealed class Result {
+        class Success(val data: String) : Result()
+        class Failure(val throwable: Throwable) : Result()
+    }
 }
